Add the array companion for hashtable (note, not exactly like the original JEMCC source, linear array instead of a linked list).  Also some cleanup on other files based on matching code and comment cleanup.

+++ src/utility/buffer.c
@@ -16,9 +16,9 @@
  *
  * @param buffer The buffer instance to be initialized.
  * @param size The number of bytes to be preallocated into the buffer.  Note
- *             that a size of zero will generate an empty buffer.
+ *             that a size of zero is allowed but will not allocate a buffer.
  * @return Reference to the internal buffer if successfully allocated or
- *         NULL on a memory allocation failure.
+ *         NULL on a memory allocation failure (if size is non-zero).
  */
 uint8_t *WXBuffer_Init(WXBuffer *buffer, size_t size) {
     if (size > 0) {
@@ -145,7 +145,7 @@ uint8_t *WXBuffer_Append(WXBuffer *buffer, uint8_t *data, size_t length,
  */
 uint8_t *WXBuffer_AppendBuffer(WXBuffer *buffer, WXBuffer *source,
                                int consume) {
-    uint64_t len = source->length;
+    size_t len = source->length;
     if (WXBuffer_EnsureCapacity(buffer, len, consume) == NULL) return NULL;
     (void) memcpy(buffer->buffer + buffer->length, source->buffer, len);
     buffer->length += len;
